requirements.txt
streamlit
langchain
langchain-community
langchain-huggingface
langchain-cohere
langchain-google-genai
langchain-core
faiss-cpu
PyPDF2
python-dotenv
torch
transformers
rank_bm25
sentence-transformers
python-docx
google-generativeai